the health point of the asteroids and the spaceship are now stored as int rather than as float

+++ include/Asteroid.hpp
@@ -32,7 +32,7 @@ public:
 private:
 	bool running;
 	float rotationVelocity;
-	float hp; // Health Points
+	int hp; // Health Points
 	short size;
 	float spread;
 	thor::ResourcePtr<sf::Texture> texture;
