+++ src/Game.cpp
@@ -27,13 +27,13 @@ Game::Game() :
 		("spaceship.weaponCooldown", po::value<float>()->default_value(0.17f))
 		("spaceship.friction", po::value<float>()->default_value(100.0f))
 		("spaceship.lifes", po::value<unsigned int>()->default_value(3))
-		("spaceship.hp", po::value<float>()->default_value(1.0f))
+		("spaceship.hp", po::value<int>()->default_value(1))
 
 		("energyBullet.speed", po::value<float>()->default_value(400.0f))
 		("energyBullet.timeToLive", po::value<float>()->default_value(0.85f))
 		("energyBullet.damage", po::value<float>()->default_value(1.0f))
 
-		("asteroid.hp", po::value<float>()->default_value(1.0f))
+		("asteroid.hp", po::value<int>()->default_value(1))
 		("asteroid.spread", po::value<float>()->default_value(25.0f))
 		("asteroid.minSpeed", po::value<float>()->default_value(50.0f))
 		("asteroid.maxSpeed", po::value<float>()->default_value(150.0f))
